Move view export logic to action
