+++ src/components/sections/people/PersonViewsPane.jsx
@@ -12,6 +12,7 @@ import { retrieveQueries } from '../../../actions/query';
 import {
     addPersonViewRow,
     createPersonView,
+    exportPersonView,
     retrievePersonView,
     retrievePersonViewColumns,
     retrievePersonViewQuery,
@@ -195,6 +196,7 @@ export default class PersonViewsPane extends RootPaneBase {
                             rowList={ rowList }
                             placeholder={ placeholder }
                             showAddSection={ this.state.viewMode == 'saved' }
+                            onDownload={ this.onClickDownload.bind(this) }
                             onPersonAdd={ person => this.props.dispatch(addPersonViewRow(viewId, person.id)) }
                             />
                     </div>
@@ -232,6 +234,11 @@ export default class PersonViewsPane extends RootPaneBase {
         else if (this.props.views.viewList.isPending) {
             return <LoadingIndicator/>;
         }
+    }
+
+    onClickDownload() {
+        const viewId = this.getParam(0);
+        this.props.dispatch(exportPersonView(viewId));
     }
 
     onClickNew() {
